Lay groundwork to integrate many applications
It wasn't correctly integrated before because, among other things,
routes were being handled inside of ReaderApp. Now the root app
takes care of everything common and the apps only of their
specific features.

ReaderApp seems to work as expected, and now it's just a matter of
writing and adding a NarratorApp.