Allow logged-in users to read unpublished chapters
This is a kind of a workaround for the preview feature. Instead of
trying to write a different preview feature, which would have to
be kept in sync and maybe would lose fidelity, allow anyone logged
in (right now only admins can do that) to read the chapter as any
chapter participant, even if the chapter is unpublished.