Workaround for the jumping cursor bug
Especially on mobile platforms there's a bug with textareas: when
editing text *in the middle* (not the end) of a textarea, the
cursor will sometimes jump to the end of the textarea.

It seems to be a bug in Elm, but there's a workaround. Using
"defaultValue" instead of "value" will reportedly make it work.