Make the media uploader more generic
The current media uploader API endpoint has a problem: namely, that
the file categorisation is automatic based on the filename, so
there's no way to tell the difference between a background image
and an image intended to be used inside a story text.

So, move towards a model where the client specifies the type of the
image itself.