Fix audio playing in mobile, Blink-based browsers
Blink seems to ignore audio playing when it doesn't start from a
user action (only in mobile platforms). See
https://bugs.chromium.org/p/chromium/issues/detail?id=138132 for
more details. Audio is started with a user action (clicking on the
"Start" button), but as it's run inside a setTimeout to introduce
a delay, the browser doesn't realise it comes from a user click.

To avoid the problem altogether, simply start the audio right away,
and only wait for the text fade-in.