Fix editor when saving a new chapter
When clicking "Save" for the first time on a new chapter, the view
changes to chapter edit. At this point the editor would be reused
and the method setDoc would be called, but it seems that method has
disappeared. So, replace that call with code that works in the new
version.