Fix the empty chapter text bug on the narrator pages
It seems that the ports were not initialized early enough, and
sometimes the Elm code would start executing, fetch data from the
backend, and try to call a port before the ports themselves were
loaded. So, tweak a bit the JavaScript loading order in the HTML
file. Not sure if there's still a race condition, but at least
it's much less likely than before.

While at it (and as a helper for debugging), add RoboHydra to the
dev dependencies.