Show new, unpublished chapters first
Strictly speaking, the current ordering uses the creation date if
not published, or the publication date if published. That means
that chapters that have been created or published last, will appear
first.

This should work for most sensible workflows: if you write a
chapter but never publish it and abandon/forget about it, it will
eventually fade away.