Make breadcrumb navigation more consistent
Create a breadcrumbNavView common view to make all breadcrumb
navigation behave the same, including that its links will NOT
trigger a full page load. Modify the chapter control page so
that the narration is fetched from the network, so its name can
be shown in the breadcrumb navigation.

Also, add breadcrumb navigation to the narration overview page.