Fix ProseMirror editors now that it's a proper SPA
Now that the page is never reloaded, the caching of ProseMirror
views in the editorViews object was making the editors not work.
The reason is that, when the second page is shown that uses an
editor, the cached DOM element is not the same one in the current
DOM with the same id.

The cache is still needed because it's used in updateParticipants,
but it might be possible to solve that problem some other way and
completely get rid of the cache.