Fix bug in chapter creation
There was a state mismatch when viewing a chapter, then creating a
new one (the editor was never initialised so it was impossible to
write new chapters after reading one).

While at it, simplify the state handling for that a little bit and
make sure (with requestAnimationFrame) that the underlying DOM
element will be there when initialising an editor.