Show reactions when writing new chapters
Before, reactions where fetched for a given chapter, so when
writing a new one (when it doesn't have an id yet because it hasn't
been saved) it was impossible to do so. Instead you had to save a
first draft to fetch the reactions.

Because it was irritating, add another endpoint to fetch the last
reactions for a narration, and make the frontend used that for new,
unsaved chapters.