Add very rough, first version of "novels"
Sometimes, and especially once a narration is done, it would be
nice to be able to give someone a URL so that they can read the
whole narration from a certain character's POV, like a novel.

The feature must not expose character UUIDs, even after having
finished the narration (just in case, more than anything), so they
use their own novel UUID to make the connection to the character
(which already gives the narration and all the rest of the
information needed to show the "novel").

It's currently not possible to create new "novels", but if they are
inserted in the database by hand, they will be shown in the UI and
they will work.

About half the code is copied from ReaderApp, but it might not be
easy to remove the duplication, so keeping them completely separate
for now.