Fix an issue with widths in the chapter editing page
Due to a small CSS issue, the right column (that holds the chapter
text, as opposed to the reactions from the previous chapter) took
much more space than it should. The problem is that in CSS, it's
only "section" elements under ".two-column" that take half the
page. We could instead tweak the CSS rule, which maybe it would be
a good idea, but just in case let's stick to sections for now.