Avoid side-effects before having a valid session
Pages that need a non-anonymous session often make HTTP requests
when loaded. As things were before, these requests (or any other
side-effects) were performed parallel to session checking, which
resulted in useless requests that got 401 as response.

Not only these requests are useless and it's sloppy to make them,
but some pages (eg. the chapter edit page) showed a 401 error
*once logged in* if that was the first page to be visited.