Improve image upload dialog, dialog looks in general
When new images are uploaded through the dialog, add them to the
dropdown and select them by default so it's clearer which image
will be inserted in the document when clicking "OK".

While at it, improve the looks (CSS) of ProseMirror dialogs in
general. This also affects the Private Text dialog.