Add a status field for narrations
Add a status field for narration so that they can be archived as
finished or abandoned. Modify the "narration overview" in the home
page to only show active narrations, and add a button to change
between narration status.

For now there's no way to find a non-active narration once it's
marked, and there's no button to mark a narration as abandoned.