Fix issue with images
Possibly after the ProseMirror upgrade the parser got more strict
and it's chocking with images used as a block, and not as an inline
element.

They were being "promoted" from inline to block with a server-side
hack so that it was possible to style images so that they took the
whole screen width, and not just the containing paragraph width.

To fix, move the hack to the client side, *only* when rendering
text, not when feeding the text to an editor (the editor is what
chokes for now). It's not an ideal solution, but it works for now
and in any case it's clear that (1) the server shouldn't mess up
with the text, at least not in a way that breaks the schema; and
(2) the solution for this problem, which is related to rendering,
should clearly be on the client side.