Add a way to create novels from the UI
This needed some amount of refactoring to make things a bit more
consistent. Unfortunately, the API endpoint URL is somewhat
inconsistent, but it's better because it's simpler, there's fewer
chances for bugs and less logical repetition (eg. basing it on the
narration id would need that the character id or token is also
provided, but then it would be possible to send a character id that
doesn't appear in the given narration).