Tweak indentation, update TODO list
After some tests, it seems that trying to figure out when the
background image and the audio will be loaded won't be that
reliable in Firefox, and the problem is that chapters might never
load in that case.

So dropping the idea for now, and move the first time configuration
idea to the top, as it's the only real showstopper left.