Add an initial setup "wizard"
Instead of shipping with a crappy default admin user with a fixed
password, make a setup screen for the first run. It will trigger
any time there isn't an admin user, so even if the only admin user
removes the admin role from themselves, it will be possible to
recover the application without going to the database.

While at it, remove old cruft that wasn't used any more.