Show only relevant "last chapters"
Although the "last reactions" were being correctly filtered by the
current chapter's participants, "last chapters" was showing the
last chapters for any potential participants. Filter them so that
only the appropriate last chapters are shown (which is a bit
trickier than with the reactions, as the markup has to be generated
anyway, then maybe hidden with CSS, due to rendering being done via
JavaScript through ports and such).

While at it, reformat the code in ChapterEditApp.Views.