Add a configuration setting to help test e-mail
When developing with a database which might be copied from a
production system, it's useful to be able to force all e-mail to be
sent to a given address, regardless of the address it's meant to be
sent to.

This avoids sending e-mails about internal, local tests to random
people.