Model chapter background image and audio better
For the reading applications (ReaderApp and NovelReaderApp) the
model of background image and audio was wrong. It always expected
a string, when it could be null.

So, model both fields correctly and adapt the markup to actual
reality, including that, when reading a single chapter (not novel),
it doesn't make sense to even show the "Background audio" checkbox
on the start page.