Fix music playing in novels; improve fade-ins
Mobile Blink (Chrome, Opera, ...) will only play audio when
initiated by a user action. That means that setTimeout or
requestAnimationFrame cannot be used to start the audio (even if
that setTimeout/requestAnimationFrame were called through a user
action).

When switching chapters in novels, it's impossible to make sure
the right audio file is already loaded. However, as we know all
audio URLs from the start, simply create as many audio elements as
chapters, and then simply switch which one is playing.

While at it, make the chapter fade-ins more reliable, by applying
the CSS transition only when fading in.