Improve formatting of message notification recipients
Streamline the formatting of the recipient list in message
notifications. Before, messages sent by the narrator would say
that the narrator sent a message to themselves and whoever else.
Now all formatting (both in messages to players and messages to
the narrator) is handled by a new function.

The small downside is that now the narrator doesn't appear in the
recipient list in notifications for players, wrongly hinting that
the narrator cannot read the messages when they are for other
players. But that's better than the previous mess, and handling
that case correctly would make the code uglier.

Add initial tests for the message notifications in the Mailer
class. As this class also depends on the database, a refactoring
was needed to make both this and NarrowsStore tests run properly.