Fix display of images, including mentions
Images in rendered chapters were too wide and displayed outside of
the containing box in eg. the control panel. Fix it so that images
take the full width of the containing box.

While at it, show which images have mentions by adding a border
with the same colour gradient as text mentions.