Init new chapter editors with *all* participants
For some reason, editors for new chapters were initialised with no
participants, and thus it wasn't possible to mark any text as
private before clicking on "Save" for the first time (which happens
to trigger another editor initialisation, this time with the
chapter's actual participants).

Instead, editors for new chapters should be initialised with *all*
characters defined in the narration, because that's the default
for new chapters.