Revert "Fix the jumping cursor bug in the affected text boxes"
The problem with using defaultValue is that it's only set when the
textarea is created. Thus, it's impossible to eg. reset the message
text to an empty value when a message is sent. It's most likely a
bug in Elm, and that bug is better than not having control over the
contents of the textarea, so it's better to revert this fix.

This reverts commit 728e23452162179aa854861c33ae1e3b6e56fa86.