Add narration edit screen to set default media
For that, refactor media upload button into fileSelector and make
it support different "targets". The function signature becomes
pretty nasty though, but that's better than repeating the upload
button code over and over. Afterwards it might be possible to
improve the function signature by passing an object of a new type
instead of many parameters.