Add Docker configuration files
Add Dockerfile and docker-compose.yaml to quickly get a local copy
running. That configuration can also serve to host a production
environment with modifications.

As sendmail and Docker don't get along, upgrade nodemailer and add
mail settings to the configuration file so that mail sending
doesn't have to be via sendmail.

BECAUSE OF THE UPGRADED NODEMAILER, NOW NARROWS REQUIRES AT LEAST
NODE 6.x!