Fix structure of empty chapters
When creating a new chapter, if the Save or Publish buttons are
pushed without writing anything in the box, the saved chapter
will have an invalid structure (it will be an empty array,
instead of an object with the "type" and "content" properties).
To fix the issue: define the default, empty chapter correctly.