Make the database connection more robust
Instead of connecting to the database directly and keeping a single
connection, use a connection pool. This means that when the
database is restarted (eg. when upgraded) the application will keep
working as normal, instead of crashing and not working again until
restarted.

However, now it's possible to have unhandled exceptions (when the
database is actually down) so create a simple, static error page
for now.

Closes #6.