Drop reactions and use messages instead
Having both reactions and messages is not very useful: one extra
concept, and requires players to type their actions twice if they
want to collaborate meaningfully. So, better to only have messages
and have their "actions" be public by default.

What becomes a bit worse is having *part* of the action secret, as
it now requires a second message (one for the public part with all
participants, and one for the private part with whatever recipients
are appropriate). That's an ok price to pay, considering secret
actions are not that common.

This is a very big commit, but it's hard to meaningfully break into
chunks.