Fix chapter rendering issue when using the back button
As the state was never reset to "Loader" when loading the reader
UI, it was possible to reach the page while keeping the "Narrating"
state, eg. when (1) loading a chapter, (2) clicking on a character
sheet, then (3) hitting the back button or clicking on another
chapter from the list.

If this combination happened, the chapter text would never be
rendered and the page would display a background image with the
text "Chapter contents go here".

While at it, reformat the whole app/ReaderApp/Update.elm.