Create a new CharacterEditApp for narrators
When clicking on a character from the list in the narration
overview screen, send the narrator to a new character edit screen.
It's similar to the character screen for the player, but this one
is visually integrated with the rest of the narrator UI, and it's
possible to see and modify who the player is.

For that it was necessary to expose the player's e-mail in the
character API.