Return the intro sent field for chapter participants
The API returning the information for a given chapter wasn't
returning the intro_sent field for the chapter participants. In
itself it doesn't have to be a big problem, but it's somewhat
ugly, plus it triggered a bug in a comparison made in the chapter
edit interface, which broke the chapter participant list.