First implementation of narration chapter search
Add a way to search for a word/phrase in the chapters of a given
narration. The logic of this is that, when writing new chapters,
it's hard to remember what has happened in other chapters, or when,
or whether a character has been introduced or how it was presented.

Expose that functionality in the chapter edit screen.