Add introductions to narrations
Add a special chapter to narrations (an intro of sorts that can be
used as a description for the story, themes, and setting) that is
always visible to anyone who has a special URL. That page will show
the characters for the narration, and will allow anyone to "claim"
a character and assign it to themselves. Show this intro, if
present, at the start of the narration "novels".

Remake the whole narration edit page to have the intro text and
other fields, and use the space a little better.

This replaces the awkward, offline way NARROWS used to work before
when starting a narration, in which the description for the story
had to be sent through other means, and the decision about who took
each character was also offline, and then the narrator had to assign
the characters by hand. Now people can do it themselves.

To make this work, give all narrations a token that is used in the
narration intro URL. Anyone who knows this URL can read the
narration intro and claim a character, but at the same time this
opens the possibility to hide narrations from the general public
(as the URLs are not guessable) if in the future there is a
directory of open narrations in search of players.

While at it, fix default/empty intro values: an empty array is
actually NOT a valid ProseMirror document. Instead, use an actual
NULL as a default/empty ProseMirror document value.

Also while at it, remove all of the intro email stuff: now that we
are changing the model to something based on people claiming
characters, it's better to simplify the application a bit and
remove all the code related to sending intro emails.