Small refactoring of backend ProseMirror text handling
Make sure that all special handling of ProseMirror data on the backend
(converting from JS object to text, so it can saved into the database;
and converting from text to JS object, so it can be returned by the
API) is done inside the NarrowsStore class, and never in the endpoints
file.

The endpoints file should always deal only with objects, and the store
should do all the necessary conversions to make that the case.