Make email optional
The email field used to be mandatory. However, in preparation for
the upcoming changes for the narration intro and players claiming
characters themselves, it's much better if characters are not
assumed to be always taken by a player, and instead are claimed or
unclaimed.

As part of this change, make sure that email notifications don't
choke on unclaimed characters, ie. characters that don't have any
associated player, and thus no associated email address. This
affected the chapterPublished and messagePosted email notifications.