Fix password reset functionality
The password reset functionality was broken for non-admin users,
due to a silly JSON response format expectation mismatch on the
frontend. In particular, the "role" field for a user is supposed to
be a string, but it was null for non-admin users. So force the role
field to be a string when returning user information, the same way
we already do when we get the list of users.