Show error when checking session
Before, any error in the session checking request was taking to be
a sign that the user was anonymous, without further checks. This is
very confusing when there are bugs in the session checking code,
and it returns 500, or JSON that isn't accepted by the frontend, or
whatever.

So, show error messages when there are actual errors, and only
accept the session as anonymous when the problem is that the error
code is 404.y