Always use default media in new chapters
Due to an optimisation when creating new chapters (the narrations
were only fetched if their id was not the same as the narration
currently in the store), it was possible to change the default
background image or audio for a narration, then create a new
chapter (without reloading the page), and have the new chapter not
use the latest changes in default media.

So, remove the optimisation and always fetch the narration when
creating a new chapter, to make sure the up-to-date default media
for the narration is always used.