Tweak narration status ribbon colours
Before, all narration status ribbons looked the same. Thus it was
hard to tell at a glance which ones were abandoned and which ones
were finished. So, assigned a different colour to the abandoned
ones (a sepia-like colour).

While at it, use bold font for the letters on the ribbon.