Trivial CSS clean-ups
Remove unused CSS class "page-aside": it was "apparently" wrongly
referenced, and thus never applied, so it wasn't making any
difference. So, better to remove anything that is not needed to
avoid confusion and ease maintenance.

Also drop the "no-content" CSS class, as it's unused now, and move
the narration ribbon styles closer together.