Improve chapter lists
Instead of using a plain unordered list to show a list of chapters,
show a more standard table-like list with alternating background
colours.

For that, the way extra chapter information (like number of messages
and missing reactions) is shown had to change a bit.

Finally, separate between "compact" chapter list view, as shown on
the frontpage; and regular chapter list view, as shown on the
narration overview page.