Improve styling of message recipients
Label styling was a bit of a mess in the application. There was a
blanket CSS rule for all labels that in practical terms it was only
used in the selection of message recipients (both for narrators and
for players).

Remove the blanket label rule and improve styling for recipient
selection.