Better fix for the character sheet white bottom
When there isn't enough content to make the character sheet screen
scrollable, we end up with an ugly white bottom bar (the page
generally has a black background). Instead of trying to fix it with
min-height 100% and (presumably) some margin black magic, use
min-height 100vh. It is more explicit and seems to work better.