Fix some Firefox layout issue with avatars
For some reason, even though the CSS states that character avatars
should be 100x100, Firefox often fucks it up and renders narrower
images. When right-clicking on the image and examining it, it
corrects itself to be 100x100, though.

It might have to do with Flex in some way, but for good measure
let's just add width and height attributes to the img tags in all
cases. This seems to fix the issue and always display the images
correctly.