Fix problem with ProseMirror dirty flag checks
It seems there is a bug in Elm when comparing Json.Encode.Value
values, and it triggers when checking if the ProseMirror texts had
really changed. This check was needed because the ProseMirror-based
editor was sending updates to Elm every time the document state
changed, including selection/cursor position changes.

So, as I don't see any way to fix the Json.Encode.Value comparison in
Elm, instead compare the documents in JavaScript (only the actual
document contents, not the selection) and only send updates to Elm
when the document contents have changed. That way we don't need to
compare anything on the Elm side, and we can assume that when we
receive a notification from JavaScript it means that the document has
changed.