First step towards removing email from view
We don't want to be showing the players' email if we want to allow
NARROWS to be used by many people who don't know each other (for
privacy/legal reasons). Now that we have narration intros and players
can claim character themselves, there's no need for the narrator to
know or be able to change the email of the player playing a given
character.

* Add a "display_name" field to the users table in the database
* Remove email field from character creation screen
* Show the player's display name instead of the email on the character
  edit screen
* Calculate which characters are taken by using the display name
  field, not the email field
* Replace email with displayName where possible in backend, when
  providing information about which users play which characters