Implement simple user deletion
For this to work, the foreign key in characters (played_id field) had
to be changed so that the field would be set to NULL when deleting a
user. We really don't want to delete the characters a user claimed
only because the user gets deleted, and we don't want to set the field
to NULL by hand if the database can take care of that.