Check permissions on all endpoints
Several endpoints missed proper permission checks, and it was possible
to fetch certain data or perform certain operations without having
proper permissions, and in some cases without even logging in. Yikes.

While at it, make userStore.isAdmin a bit more reliable, and reorder
the backend routes to make it a bit easier to scan and see which ones
should have checks, and what kind.

Finally, drop the function apiAdminAuth: it's not needed now that
we're explicitly checking _inside the endpoint function_ if the user
is an admin, and in general it's better to make all of the checks in a
single place, and that place is the endpoint definition.