Fix character claim emails
Since the code removed the "email" field from most return values, two
things broke in character claim emails:

1. The "Now all character in your story are claimed" was always there
2. The fellow players were not notified when a new character was
claimed

So, move some code around and make sure we have the email addresses
available so that the character claim emails are sent to everyone, and
with the correct contents.

Also, there was something broken even before the display name change:
the name of the claimed character was wrong in the fellow player
emails! So, fix it, too.