Use character colours for their preview buttons
When editing a chapter, there is a button per participant character
that shows the chapter preview from that character's perspective.
Because we already assign a colour for the chapter fragments available
for that character, it makes sense to use the same colours for the
character preview buttons so they are easier to identify.