Fix textbox width issue in Firefox
The textboxes on the right column in the character edit screen had the
wrong width, pushing the buttons they had besides ("Unclaim" and
"Reset") out of the column width. This only happened on Firefox.

As it turns out, Firefox was getting confused because the textboxes
had both a "size" attribute in HTML, and a "width" CSS property. So,
remove the size in those textboxes so everything looks as it should.