Create a new getNarratorId method in NarrowsStore
Most uses of getNarration are only to get the narrator id. Having a
specific method that only returns that piece of data is better because
that way it's clearer that the id is the value used. It also makes the
code easier to refactor because we don't have to worry about the
frontend being able to parse the changes everywhere, when we add or
change fields in the return value for getNarration.