Redo character avatars completely
Create a function to render character avatars. Apart from having that
code in a single place, this new function also creates a better
default when there isn't an avatar (using the first two initials of
the character's name), sets alt on the images, and always sets title
on the avatar so the full name is always available.