Make character sheet look like the rest of the app
Instead of having the character sheet app use a dark background like
the chapter reader app, drop the "reader-app" id from the <main>
element and make it use whatever colours are used for the rest of the
app.

There's a happy side-effect to this, namely that "reader-app" becomes
a marker for the chapter-reading apps: for regular chapters, for the
novel, and for the narration intros.