Make sure avatars have the right width
In the Dramatis personae section of chapters, placeholder avatars
(the black circle with the character initials) didn't quite take
the full width they should. This is due to Flexbox taking some
width from that element to give it to the character description
besides it.

So, force the avatar to never shrink as a Flexbox item so that it
always uses that full 100px width.