Fix claim email text
Don't talk about "writing the first chapter" in the character claim
emails if there are already published chapters in that
narration. Normally the characters are claimed before starting, but
eg. when a character is abandoned, a character can be claimed
mid-story. This is awkward because the email will say that you should
wait until the narrator writes the first chapter, which is something
that has already happened potentially long ago.

While at it: catch errors when sending unclaim emails, and fix the
plain text version of one of the emails, which started with a broken
HTML element.